--- src/Modal.js
@@ -1,23 +1,0 @@
-import { useEffect, useRef } from 'react';
-import { createPortal } from 'react-dom';
-
-/**
- * @param {React.Children} children
- * @return {React.ReactPortal}
- */
-export const Modal = ({ children }) => {
-  const elementRef = useRef(null);
-
-  if (!elementRef.current) {
-    elementRef.current = document.createElement('div');
-  }
-
-  useEffect(() => {
-    const modalRoot = document.getElementById('modal');
-    modalRoot.appendChild(elementRef.current);
-
-    return () => modalRoot.removeChild(elementRef.current);
-  }, []);
-
-  return createPortal(<div>{children}</div>, elementRef.current);
-};
